+++ cmd/xgoja/internal/buildspec/spec.go
@@ -60,6 +60,8 @@ func (m ModuleInstance) Ref() string {
 
 type CommandsSpec struct {
 	Repl    CommandSpec `yaml:"repl" json:"repl"`
+	Run     CommandSpec `yaml:"run" json:"run"`
+	TUI     CommandSpec `yaml:"tui" json:"tui"`
 	JSVerbs CommandSpec `yaml:"jsverbs" json:"jsverbs"`
 }